[clang][bytecode] Move shard interp functions to InterpHelpers.h
This drastically reduces the preprocessed size of Context.cpp and
InterpBuiltin.cpp.
